server/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
phonenumbers==8.13.50
psycopg2-binary==2.9.10
PyJWT==2.10.0
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
telebot==0.0.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
